requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
WTForms==3.1.2
Pillow==10.4.0
python-dotenv==1.0.1
Werkzeug==3.0.3
SQLAlchemy==2.0.31
